+++ hook.py
@@ -2,7 +2,8 @@ import datetime
 from time import sleep
 from time import gmtime, strftime
 
-from test_code.lcd_1602 import LCD_LINE_1, LCD_LINE_2, lcd_string
+from lib.lcd import lcd_string
+import lib.lcd as lcd
 
 def close_door_hook():
     show_datetime()
@@ -12,6 +13,6 @@ def open_door_hook():
 
 def show_datetime(): #「日時」を表示
     local_time = datetime.datetime.now()
-    lcd_string(strftime("%Y.%m.%d (%a)", gmtime()) , LCD_LINE_1)
-    lcd_string(local_time.strftime("%H:%M"), LCD_LINE_2)
+    lcd_string(strftime("%Y.%m.%d (%a)", gmtime()) , lcd.LCD_LINE_1)
+    lcd_string(local_time.strftime("%H:%M"), lcd.LCD_LINE_2)
     sleep(1)
