+++ main.py
@@ -6,81 +6,12 @@ import pigpio
 from time import sleep
 import time
 import datetime
-import smbus
 import pandas as pd
 import datetime
 
+import lib.lcd as lcd
+from lib.lcd import lcd_byte, lcd_string, lcd_init
 from hook import open_door_hook, close_door_hook
-
-# Define some device parameters
-I2C_ADDR  = 0x27 # I2C device address 
-LCD_WIDTH = 16   # Maximum characters per line
-
-# Define some device constants
-LCD_CHR = 1 # Mode - Sending data
-LCD_CMD = 0 # Mode - Sending command
-
-LCD_LINE_1 = 0x80 # LCD RAM address for the 1st line
-LCD_LINE_2 = 0xC0 # LCD RAM address for the 2nd line
-
-LCD_BACKLIGHT  = 0x08  # On
-#LCD_BACKLIGHT = 0x00  # Off
-
-ENABLE = 0b00000100 # Enable bit
-
-# Timing constants
-E_PULSE = 0.0005
-E_DELAY = 0.0005
-
-#Open I2C interface
-bus = smbus.SMBus(1) # Rev 2 Pi uses 1
-
-
-### LCD control
-
-def lcd_init():
-  # Initialise display
-  lcd_byte(0x33,LCD_CMD) # 110011 Initialise
-  lcd_byte(0x32,LCD_CMD) # 110010 Initialise
-  lcd_byte(0x06,LCD_CMD) # 000110 Cursor move direction
-  lcd_byte(0x0C,LCD_CMD) # 001100 Display On,Cursor Off, Blink Off 
-  lcd_byte(0x28,LCD_CMD) # 101000 Data length, number of lines, font size
-  lcd_byte(0x01,LCD_CMD) # 000001 Clear display
-  sleep(E_DELAY)
-
-def lcd_byte(bits, mode):
-  # Send byte to data pins
-  # bits = the data
-  # mode = 1 for data
-  #        0 for command
-  global LCD_BACKLIGHT
-  bits_high = mode | (bits & 0xF0) | LCD_BACKLIGHT
-  bits_low = mode | ((bits<<4) & 0xF0) | LCD_BACKLIGHT
-
-  # High bits
-  bus.write_byte(I2C_ADDR, bits_high)
-  lcd_toggle_enable(bits_high)
-
-  # Low bits
-  bus.write_byte(I2C_ADDR, bits_low)
-  lcd_toggle_enable(bits_low)
-
-def lcd_toggle_enable(bits):
-  # Toggle enable
-  sleep(E_DELAY)
-  bus.write_byte(I2C_ADDR, (bits | ENABLE))
-  sleep(E_PULSE)
-  bus.write_byte(I2C_ADDR,(bits & ~ENABLE))
-  sleep(E_DELAY)
-  
-
-def lcd_string(message,line):
-  # Send string to display
-  message = message.ljust(LCD_WIDTH," ")
-  lcd_byte(line, LCD_CMD)
-  for i in range(LCD_WIDTH):
-    lcd_byte(ord(message[i]),LCD_CHR)
-
 
 class Servo:
     #set servo PIN
@@ -174,12 +105,12 @@ def main():
     lcd_init()
     servo.door_open() ##最初は開けておく
     
-    lcd_string("   Welcome to   ", LCD_LINE_1)
-    lcd_string(" CAMPHOR- LOCK ", LCD_LINE_2)
+    lcd_string("   Welcome to   ", lcd.LCD_LINE_1)
+    lcd_string(" CAMPHOR- LOCK ", lcd.LCD_LINE_2)
 
     sleep(1) #LCDを1秒間点灯
     LCD_BACKLIGHT = 0x00  #バックライトオフ
-    lcd_byte(0x01, LCD_CMD) #表示内容クリア
+    lcd_byte(0x01, lcd.LCD_CMD) #表示内容クリア
 
     while True:
         ret = card.nfc_process()
@@ -187,50 +118,50 @@ def main():
             card_istouch = True
             card_touched_time = time.time()
             LCD_BACKLIGHT  = 0x08  # On
-            lcd_byte(0x01, LCD_CMD) #表示内容クリア
+            lcd_byte(0x01, lcd.LCD_CMD) #表示内容クリア
             is_timeout = False
 
             if nfc_data.check_registered(ret): #idmがすでに登録されているカードだったら
                 if want_close == True and door_isopen == True:
-                    lcd_string(" GOOD BYE >_< ", LCD_LINE_1)
-                    lcd_string(" SEE YOU AGAIN ", LCD_LINE_2)
+                    lcd_string(" GOOD BYE >_< ", lcd.LCD_LINE_1)
+                    lcd_string(" SEE YOU AGAIN ", lcd.LCD_LINE_2)
                     servo.door_close()
                     door_isopen = False
                     want_close = False
 
                 elif want_close == False and door_isopen == False:
-                    lcd_string("   Welcome to   ", LCD_LINE_1)
-                    lcd_string(" CAMPHOR- HOUSE ", LCD_LINE_2)
+                    lcd_string("   Welcome to   ", lcd.LCD_LINE_1)
+                    lcd_string(" CAMPHOR- HOUSE ", lcd.LCD_LINE_2)
                     servo.door_open()
                     door_isopen = True
                     
             else: #登録されていないカードだったら
-                lcd_string("  THIS CARD IS  ", LCD_LINE_1)
-                lcd_string(" NOT REGISTERED ", LCD_LINE_2)
+                lcd_string("  THIS CARD IS  ", lcd.LCD_LINE_1)
+                lcd_string(" NOT REGISTERED ", lcd.LCD_LINE_2)
                 while True: #5秒間ボタン入力を受け付ける
                     if servo.switch_status(): #ボタンの入力がされたら
-                        lcd_string(" PLEASE TOUCH  ", LCD_LINE_1)
-                        lcd_string(" REGISTERED CARD", LCD_LINE_2)
+                        lcd_string(" PLEASE TOUCH  ", lcd.LCD_LINE_1)
+                        lcd_string(" REGISTERED CARD", lcd.LCD_LINE_2)
                         wait_master_time = time.time()
 
                         while True: #すでに登録されたカードのタッチを15秒待つ
                             ret_m = card.nfc_process()
                             if time.time() - wait_master_time > 15:
                                 is_timeout = True
-                                lcd_string("REGISTERATION IS", LCD_LINE_1)
-                                lcd_string(" TIMEOUT... >_< ", LCD_LINE_2)
+                                lcd_string("REGISTERATION IS", lcd.LCD_LINE_1)
+                                lcd_string(" TIMEOUT... >_< ", lcd.LCD_LINE_2)
                                 break
                             if ret_m is not None:
                                 if nfc_data.check_registered(ret_m): #idmがすでに登録されているカードだったら
                                     nfc_data.registering_card(ret) #元々タッチされたカードを追加する
-                                    lcd_string(" THIS CARD IS  ", LCD_LINE_1)
-                                    lcd_string("REGISTERED! ^o^ ", LCD_LINE_2)
+                                    lcd_string(" THIS CARD IS  ", lcd.LCD_LINE_1)
+                                    lcd_string("REGISTERED! ^o^ ", lcd.LCD_LINE_2)
                                     sleep(1)
                                     is_timeout = True #ループから抜けるために便宜上
                                     break
                                 else:
-                                    lcd_string(" THIS CARD IS  ", LCD_LINE_1)
-                                    lcd_string(" NOT MASTER CARD ", LCD_LINE_2)
+                                    lcd_string(" THIS CARD IS  ", lcd.LCD_LINE_1)
+                                    lcd_string(" NOT MASTER CARD ", lcd.LCD_LINE_2)
   
                     if time.time() - card_touched_time > 5:
                         is_timeout = True
@@ -243,14 +174,14 @@ def main():
                 #LCDの焼け付きを防止するために消す
                 card_istouch = False
                 LCD_BACKLIGHT = 0x00  #バックライトオフ
-                lcd_byte(0x01, LCD_CMD) #表示内容クリア
+                lcd_byte(0x01, lcd.LCD_CMD) #表示内容クリア
 
         if servo.switch_status():
             if want_close != True:
                 LCD_BACKLIGHT  = 0x08  # On
-                lcd_byte(0x01, LCD_CMD) #表示内容クリア
-                lcd_string("HOUSE CLOSE MODE", LCD_LINE_1)
-                lcd_string(" TOUCH KEY ^o^ ", LCD_LINE_2)
+                lcd_byte(0x01, lcd.LCD_CMD) #表示内容クリア
+                lcd_string("HOUSE CLOSE MODE", lcd.LCD_LINE_1)
+                lcd_string(" TOUCH KEY ^o^ ", lcd.LCD_LINE_2)
                 want_close_time = time.time()
                 want_close = True
 
@@ -259,7 +190,7 @@ def main():
                 #LCDの焼け付きを防止するために消す
                 want_close = False
                 LCD_BACKLIGHT = 0x00  #バックライトオフ
-                lcd_byte(0x01, LCD_CMD) #表示内容クリア    
+                lcd_byte(0x01, lcd.LCD_CMD) #表示内容クリア
